Added refresh for viewSQL field in multi-instance mode (#6895)
Currently the refresh callbacks for distributed locks in multi-instance mode refresh Table Descriptors from the mapd_tables table in the catalog, however, this table does not contain some of the needed information for Views (namely the viewSQL field in the TD). This data is contained in the mapd_views table and needs to be refreshed separately.

This PR adds refreshing capability for this field in the Catalog.

Testing has been done separately and is included in a follow-up PR. Prior to this PR, View refreshing in multi-instance was broken and had no associated testing.

Signed-off-by: jack <jack@heavy.ai>